updated file structire. Home, Login, and GymLeaderHome are now all pages. Authenticate, collection, gym-leader, import, and profile are features. App.js routes to the various pages
