+++ packages/client/src/features/gym-leader/CardInfo.js
